requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
aiofiles>=23.2.1
requests>=2.31.0
clickhouse-connect>=0.7.0
python-dateutil>=2.8.2